Rename mapper and graph views
