+++ frontend/src/App.js
@@ -7,14 +7,14 @@ import { useLoginStateValue } from './lib/store'
 import NavBar from './components/NavBar'
 import WorkspaceNavBar from './components/WorkspaceNavBar'
 import PrivateRoute from './components/PrivateRoute'
-import CourseMapperView from './views/mapper/CourseMapperView'
+import CourseMapperView from './views/coursemapper/CourseMapperView'
 import HomeView from './views/home/HomeView'
-import MapperRedirectView from './views/mapper/redirect/MapperRedirectView'
+import MapperRedirectView from './views/coursemapper/redirect/MapperRedirectView'
 import WorkspaceManagementView from './views/manager/WorkspaceManagementView'
 import JoinView from './views/join/JoinView'
 import HeatmapView from './views/heatmap/HeatmapView'
 import LoginView from './views/login/LoginView'
-import CytoGraphView from './views/graph/CytoGraphView'
+import CytoGraphView from './views/graph/GraphView'
 import ProjectView from './views/project/ProjectView'
 import ProjectNavBar from './components/ProjectNavBar'
 import CloneView from './views/project/CloneView'
