+++ frontend/src/views/coursemapper/ActiveCourse.js
