+++ frontend/src/views/coursemapper/Course.js
