+++ frontend/src/views/coursemapper/CourseContainer.js
