+++ frontend/src/views/coursemapper/CourseMapperView.js
