+++ frontend/src/views/coursemapper/MapperLinks.js
