+++ frontend/src/views/coursemapper/redirect/MapperRedirectView.js
